chatbot/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
dialogflow
python-dotenv
aiohttp
botbuilder-core
requests
jose